redux thunk middleware added
