+++ package.json
@@ -19,6 +19,7 @@
     "react-toastify": "^9.1.1",
     "react-uuid": "^2.0.0",
     "redux": "^4.2.0",
+    "redux-devtools-extension": "^2.13.9",
     "redux-thunk": "^2.4.2",
     "web-vitals": "^2.1.4"
   },
