--- src/Redux/actions/Types.js
@@ -1,2 +1,0 @@
-export const GET_COMPANY = 'GET_COMPANY'
-export const GET_COMPANY_ERROR = 'GET_COMPANY_ERROR'
